requirements.txt
Flask==1.0.2
pyBarcode==0.7
Pillow==5.2.0